msm: cpr: Add copy from smem for floor_fuse bits
Copy the new floor_fuse bits from smem to app. Update the
padding in the cpr smem structure. As per the new addition
of floor_fuse bits, it is required to add the same in the
structure used by app and thereby copy the value from smem
into the local copy of app. The padding is reduced by 1 byte
since the floor_fuse is of 1 byte. The debug print has been
changed to pr_info for those vital information that are needed
for debug purpose and to assess which type of part the CPR was
running.

Change-Id: Ia1f9dad6519b18360ad7e81c765cc85670ac8708
Signed-off-by: Kaushal Kumar <kaushalk@codeaurora.org>
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>

Conflicts:

	arch/arm/mach-msm/devices-msm7x27a.c