msm_fb: board: Set MIPI lane swap register in the board file
Currently the MIPI lane swap register is set in the panel file.
But setting the register depends on the controller lanes that come
out of the MSM and also the way in which these controller lanes
are connected to the LCD connector. Hence, move this setting
to the board file so that register setting is done based
on the MSM chipset and also the MSM to LCD controller lane
mapping configuration on the respective board.

Change-Id: Ic2d98560e9451e8f071e86705c0b15b5d73236a5
Signed-off-by: Padmanabhan Komanduru <pkomandu@codeaurora.org>

Conflicts:

	arch/arm/mach-msm/board-8930-display.c