msm: cpr: Adjust step quot range as per voltage step
The max and min limits of step quot depend on the voltage
step size used. For PMIC based regulator, the voltage
step size is 12.5mV. However, for external buck
regulator, the step size is 6.25mV. Accordingly, the
step quot limits need to be divided by the same factor
i.e. 2.

Change-Id: If8590284416f3abff895af9d79250dfc568b4eae
Signed-off-by: Kaushal Kumar <kaushalk@codeaurora.org>