msm: acpuclock-8625q: Store previous freq & voltage
As of now, it is possible to retrieve the present
frequency before the crash from the dump but not
previous frequency and its corresponding voltage.
Now previous frequency is available in a11clk_khz and
previous voltage is available in vdd, both of these
are members of drv_state.prev_speed symbol in crash
dump.

Change-Id: I86ec46287b462f0ab8bedae578cfbbe878ae9703
Signed-off-by: Kaushal Kumar <kaushalk@codeaurora.org>