regulator: fan53555: API to configure the regulator during reboot
In cases when the device is not completely turned off (reboot)
the regulator registers are not reset. This can cause the device to
boot at the same voltage at which it went down, leading to
instability. Add an API to set the default voltage.

Also, call this API before the APPS goes for a restart on 8x25q.

Change-Id: Ibb388134afbb852e2fa52bfb18c2b53408cfbbc2
Signed-off-by: Tirupathi Reddy <tirupath@codeaurora.org>